It mostly works
- It's not pretty
- It can't handle infinite loops or deep recursions (and the errors are
- ugly)
- It's pretty easy to make the page crash
- There's not explanation for what's what
- And probably a lot of other stuff

But, it mostly works. There are different inputs for editing the
expression as a Lambda, a Combinator, or a Jot expression, or even as
just a number. The page also uses some rough Church encoding to "decode"
the expression as booleans, numbers, arrays, or strings. It's pointless,
but kinda cool.